refactor: update import paths for helpers in tests and hooks
